requirements.txt
Django==1.9.7
PyYAML==3.11
argparse==1.2.1
coverage==4.1
dj-database-url==0.4.1
dj-static==0.0.6
django-autoslug==1.9.3
django-extensions==1.6.7
django-rest-swagger==0.3.8
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
futures==3.0.5
gunicorn==19.6.0
psycopg2==2.6.1
six==1.10.0
static3==0.7.0
whitenoise==3.2
wsgiref==0.1.2
